feat(frontend): refactor sidebar and add statistic page
- Refactor sidebar component to improve menu structure and styling
- Add new statistic page for parking lot data analysis
- Update admin dashboard to remove redundant functionality
- Improve responsiveness and accessibility of sidebar
